bm2_mqtt/requirements.txt
bleak~=0.14.2
pycrypto~=2.6.1
paho-mqtt~=1.6.1
